working token_controller

+++ src/service/user_service.py
@@ -7,25 +7,25 @@ from .service import Service
 from dto import UpdateUserDTO
 
 class UserService(Service):
-    def create_user(self, user: User) -> User:
-        self.db.add(user)
-        self.db.commit()
-        self.db.refresh(user)
+    async def create_user(self, user: User) -> User:
+        await self.db.add(user)
+        await self.db.commit()
+        await self.db.refresh(user)
         return user
         
-    def get_all_users(self) -> list[User]:
-        return self.db.query(User).all()
+    async def get_all_users(self) -> list[User]:
+        return await self.db.query(User).all()
     
-    def get_user_by_id(self, user_id: int) -> User:
-        user = self.db.query(User).filter(User.id == user_id).first()
+    async def get_user_by_id(self, user_id: int) -> User:
+        user = await self.db.query(User).filter(User.id == user_id).first()
         return user
     
-    def get_user_by_username(self, username: str) -> User:
-        user = self.db.query(User).filter(User.username == username).first()
+    async def get_user_by_username(self, username: str) -> User:
+        user = await self.db.query(User).filter(User.username == username).first()
         return user
     
-    def update_user(self, new_user: UpdateUserDTO) -> User:
-        existing_user = self.get_user_by_username(new_user.username)
+    async def update_user(self, new_user: UpdateUserDTO) -> User:
+        existing_user = await self.get_user_by_username(new_user.username)
         if not existing_user:
             raise AppException(ErrorCode.USER_NOT_FOUND, f"User with username {new_user.username} not found")
         
@@ -33,10 +33,14 @@ class UserService(Service):
         for field in fields_to_update:
             setattr(existing_user, field, getattr(new_user, field))
                 
-        self.db.commit()
-        self.db.refresh(existing_user)
+        await self.db.commit()
+        await self.db.refresh(existing_user)
         return existing_user
     
-    def delete_user(self, user: User) -> None:
-        self.db.delete(user)
-        self.db.commit()
+    async def delete_user(self, user: User) -> bool:
+        user = await self.get_user_by_id(user.id)
+        if not user:
+            return False
+        await self.db.delete(user)
+        await self.db.commit()
+        return True
